So close to finished and stable it's unreal.
Powerful buttons now have emergency safeties!

+++ post.php
@@ -42,11 +42,14 @@ if (isset($_POST['remove_product_submit'])) {
 }
 
 if (isset($_POST['remove_stockroom_product_submit'])) {
-    $remove_product_name = $_POST['remove_product_name'];
-    $query = "DELETE FROM product_table WHERE product_name='$remove_product_name';";
-    mysql_query($query) or die ("Cannot update database");
-    #die("$query");
-    header("location:stockroom.php");
+    if (isset($_POST['remove_stockroom_checkbox'])) {
+        $remove_product_name = $_POST['remove_product_name'];
+        $query = "DELETE FROM product_table WHERE product_name='$remove_product_name';";
+        mysql_query($query) or die ("Cannot update database");
+        header("location:stockroom.php");
+    } else {
+        header("location:stockroom.php");
+    }
 }
 
 if (isset($_POST['alter_product_submit'])) {
@@ -92,12 +95,18 @@ if (isset($_POST['vending_new_machine_submit'])) {
 
 if (isset($_POST['vending_remove_machine_submit'])) {
     //*echo '<script>if(window.confirm("Are you SURE you wish to delete the vending machine, and loose record of both it and all the products it currently contains?"));</script>';
+    if ($_POST['remove_stockroom_checkbox'] == 'true') {
 
-    $machine_to_delete = $_POST['vending_remove_machine_dropdown'];
-    //echo $machine_to_delete;
-    $query = "DELETE FROM machine_table WHERE machine_id='$machine_id';";
-    mysql_query($query) or die ("Unable to delete Vending Machine");
-    header("location:vending.php");
+
+        $machine_to_delete = $_POST['vending_remove_machine_dropdown'];
+        //echo $machine_to_delete;
+        $query = "DELETE FROM machine_table WHERE machine_id='$machine_id';";
+        mysql_query($query) or die ("Unable to delete Vending Machine");
+        header("location:vending.php");
+    } else {
+        //Do Nothing!! Don't touch anything!!
+    }
+
 }
 
 if (isset($_POST['vending_alter_machine_submit'])) {
